+++ src/HttpServer/BaseHttpHandler.java
@@ -1,10 +1,9 @@
-package API;
+package HttpServer;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
 import com.sun.net.httpserver.HttpExchange;
-import com.sun.net.httpserver.HttpHandler;
 import manager.TaskManager;
 import model.NotFoundException;
 import model.Task;
