+++ src/HttpServer/DurationTypeAdapter.java
@@ -1,4 +1,4 @@
-package API;
+package HttpServer;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
