+++ src/HttpServer/HistoryHandler.java
@@ -1,45 +1,47 @@
-package API;
+package HttpServer;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
-import manager.TaskManager;
+import manager.HistoryManager;
 
 import java.io.IOException;
 import java.time.Duration;
 import java.time.LocalDateTime;
 
-public class PriorityHandler extends BaseHttpHandler implements HttpHandler {
-    public PriorityHandler(TaskManager fileBackedTaskManager) {
-        PriorityHandler.fileBackedTaskManager = fileBackedTaskManager;
+public class HistoryHandler extends BaseHttpHandler implements HttpHandler {
+    public HistoryHandler(HistoryManager inMemoryHistoryManager) {
+        HistoryHandler.inMemoryHistoryManager = inMemoryHistoryManager;
     }
 
-    private static TaskManager fileBackedTaskManager;
+    private static HistoryManager inMemoryHistoryManager;
 
 
     @Override
     public void handle(HttpExchange exchange) throws IOException {
+
         Endpoint endpoint = getEndpoint(exchange.getRequestURI().getPath(), exchange.getRequestMethod());
 
         switch (endpoint) {
-            case PRIORITIZED: {
-                handlePriority(exchange);
+            case HISTORY: {
+                handleHistory(exchange);
                 break;
             }
             default:
                 writeResponse(exchange, "Такого эндпоинта не существует", 404);
         }
+
     }
 
-    private void handlePriority(HttpExchange exchange) throws IOException {
+    private void handleHistory(HttpExchange exchange) throws IOException {
         GsonBuilder gsonBuilder = new GsonBuilder();
         gsonBuilder.registerTypeAdapter(LocalDateTime.class, new LocalDateTimeTypeAdapter());
         gsonBuilder.registerTypeAdapter(Duration.class, new DurationTypeAdapter());
         gsonBuilder.setPrettyPrinting();
         Gson gson = gsonBuilder.create();
 
-        String tasks = gson.toJson(fileBackedTaskManager.getPrioritizedTasks());
+        String tasks = gson.toJson(inMemoryHistoryManager.getHistory());
 
         writeResponse(exchange, tasks, 200);
     }
@@ -48,9 +50,9 @@ public class PriorityHandler extends BaseHttpHandler implements HttpHandler {
         String[] pathParts = requestPath.split("/");
 
         if (requestMethod.equals("GET")) {
-            if (pathParts[1].equals("prioritized")) {
+            if (pathParts[1].equals("history")) {
                 if (pathParts.length == 2) {
-                    return Endpoint.PRIORITIZED;
+                    return Endpoint.HISTORY;
                 }
             }
         }
