+++ src/HttpServer/HttpTaskServer.java
@@ -1,4 +1,4 @@
-package API;
+package HttpServer;
 
 import com.sun.net.httpserver.HttpServer;
 import manager.HistoryManager;
